Move MockAPI to accessible location in tests
