+++ tests/commands/test_command_command.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from api import MockAPI
+from tests.api import MockAPI
 from cactusbot.commands.magic import Meta
 from cactusbot.packets import MessagePacket
