+++ tests/commands/test_quote.py
@@ -1,6 +1,6 @@
 import pytest
 
-from api import MockAPI
+from tests.api import MockAPI
 from cactusbot.commands.magic import Quote
 
 quote = Quote(MockAPI("test_token", "test_password"))
