+++ tests/commands/test_trust.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from api import MockAPI
+from tests.api import MockAPI
 from cactusbot.commands.magic.trust import _trust
 from cactusbot.packets import MessagePacket
